Fix: Improve the detail view landscape layout. Fix a bug with screen rotation in the main view. Fix bug in the settings menu. Make a change to sort order trigger a call to tmdb. Clean up code so it matches coding standard.
Change the default sort order to Popular. In the main view, the wrong menu
item was being called. There were duplicate menus defined in the
MainActivity and MainActivityFragment. Removed the menu item from
MainActivity. Adjustments to the Main view column size. Playing around
with that. Change fragment_detail.xml to a scrollView. It still isn't
scrolling, and now screen rotation causes crash. Make SettingsActivity
return information to the calling activity. This is just so changing the
sort order causes a new call to tmdb. Make screen rotation
in main view not trigger call to tmdb. In the detail view, push data onto
savedInstanceState and retrieve when doing screen rotation.