Feat: Create separate layout for tablet screens sized over 600dp. Change the use of toolbar titles to better match the gist in the assignment. Fix: The phone size layout wasn’t showing up because it decided on a width of 0. Fix: The screen size wasn’t being handled correctly because the logic to check for whether there is a right-hand detail pane was in onCreateView of MainActivityFragment, and that is called before ActivityMain.onCreate finishes, so I moved that logic into MainActivityFragement.onStart()
Add the layout-sw600dp files.
Add a blank fragment for the tablet screen details view before user has selected a movie.
Add the layout and fragment for a detail panel, to be used only in screens over 600dp.
Gradle tools had to be upgraded to 2.1.0 in build.gradle file.
Adjusted some colors to be more readable.